fix(output-processing): correct calculation of massdelta_condensation
In a previous version of CLEO used for output_v4.0 and previous, the massdelta_cond had an error factor of 1e18.
In the new version of CLEO v0.30.1, this is no longer the case.
We need to remove the factor from the script which calculates the eulerian views in oder to get correct evaporation values.