fix(output-processing): correct calculation of source term in conservation dataset.
This is the same fix as in 1804d7afa07743a356ffca1a0d613e3fa41163dc but for the create_inflow_outflow_mpi4py.py script

In a previous version of CLEO used for output_v4.0 and previous, the massdelta_cond had an error factor of 1e18.
In the new version of CLEO v0.30.1, this is no longer the case.
We need to remove the factor from the script which calculates the eulerian views in oder to get correct evaporation values.